Fix critical compiler warnings (buffer overflow, undefined behavior, pointer cast)
Fixed three categories of high-priority warnings identified by Cursor AI:

1. Buffer overflow in opna.cpp (lines 1295-1297)
   - strncat() was called with MAX_PATH instead of remaining buffer size
   - Fixed: Use MAX_PATH - strlen(buf) - 1 for safe concatenation
   - Impact: Prevents potential buffer overflow when loading rhythm samples

2. Undefined behavior in lz77d.cpp (lines 38-40, 77)
   - Sequence point violation: *dest++ = dest[x];
   - Same variable modified and referenced in single expression
   - Fixed: Use temporary variable to avoid undefined behavior
   - Impact: Ensures consistent decompression on all compilers/platforms

3. Integer-to-pointer cast warnings in Z80c.cpp (lines 125, 2352)
   - Direct cast from uint to uint8* causes warnings on 64-bit
   - Fixed: Cast via uintptr_t to suppress legitimate warnings
   - Impact: Cleaner builds on 64-bit platforms, no functional change

Remaining warnings are intentionally not fixed:
- Parentheses in macro (lz77d.cpp:10) - coding style, not a bug
- Unused variables - will be used in future implementation
- Member initialization order - no functional impact
- Enum switch cases - intentional partial handling

Build verified: All targets compile successfully with reduced warnings.

+++ src/devices/Z80c.cpp
@@ -122,7 +122,7 @@ void Z80C::SetPC(uint newpc)
 		DEBUGCOUNT(5);
 		instbase = instlim = 0;
 		instpage = (uint8*) ~0;
-		inst = (uint8*) newpc;
+		inst = (uint8*) (uintptr_t) newpc;  // Cast via uintptr_t to avoid warning on 64-bit
 		return;
 	}
 }
@@ -2349,7 +2349,7 @@ bool IFCALL Z80C::LoadStatus(const uint8* s)
 	reg = st->reg;
 	instbase = instlim = 0;
 	instpage = (uint8*) ~0;
-	inst = (uint8*) reg.pc;
+	inst = (uint8*) (uintptr_t) reg.pc;  // Cast via uintptr_t to avoid warning on 64-bit
 
 	intr = st->intr;
 	waitstate = st->wait;
